requirements.txt
einops==0.8.0
fvcore==0.1.5.post20221221
numpy==1.26.3
opencv_python==4.10.0.84
Pillow==10.2.0
selective_scan==0.0.2
timm==0.4.12
torch==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.67.0
